feat: Removed inherited widget scope
Added static get for dialogsController
