+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 1.0.0
+
+* BREAKING: FlutterEasyDialogs.of(context) method for getting dialogsController replaced with static get method based on GlobalKey. EasyDialogsController is now available in any place of the app.
 ## 0.0.2
 
 * Description changed
