+++ lib/flutter_easy_dialogs.dart
@@ -6,6 +6,5 @@ export 'src/core/dialogs/dialogs.dart';
 export 'src/core/dialogs/pre_built/easy_modal_banner/easy_modal_banner_theme_data.dart';
 export 'src/core/dismissibles/dismissibles.dart';
 export 'src/core/dismissibles/types/easy_positioned_dismissible_type.dart';
-export 'src/core/flutter_easy_dialogs/flutter_easy_dialogs.dart';
 export 'src/core/flutter_easy_dialogs/flutter_easy_dialogs_theme.dart';
 export 'src/core/overlay/overlay.dart' hide EasyOverlayAppEntry, EasyOverlay;
